Add DigiCert Global G2 and Microsoft RSA Root CA 2017
Azure will move on to new certs in 2023. As this package pins them
they need to be updated. In order to not disrupt users, we should
follow Azure's guidance and add the new certs as soon as possible.

At some point in 2023, Azure will remove the DigiCert Baltimore
Root, but for now it is recommended to keep.

See the following URL for the announcement and updates:
https://techcommunity.microsoft.com/t5/internet-of-things-blog/azure-iot-tls-critical-changes-are-almost-here-and-why-you/ba-p/2393169