Convert TwinState to a []byte.
At the MQTT layer publishing and receiving twin state is a []byte. This
eliminates the conversion of that data to map[string]interface{} at the
Client layer. I believe the original implementation expected twin data
to only be key/value pairs. This was not the case for me and an
interface{} type can not be indexed. Working with the byte array better
works in my use case.